taller_clase_5/requirements_annotations.txt
matplotlib>=3.5.0
# Dependencias para visualización de anotaciones Caltech-101
numpy>=1.21.0
Pillow>=8.3.0
scipy>=1.7.0
